// hgmall-frontend/src/api/generated/models/shop-vo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TheMall API 文档
 * TheMall 商城系统 API 接口文档
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@themall.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ShopVO {
    'shopId'?: number;
    'merchantId'?: number;
    'name'?: string;
    'status'?: number;
    'description'?: string;
    'createTime'?: string;
    'updateTime'?: string;
}